set up user repository

+++ Back End/FindMyStuff/src/main/java/com/example/demo/database/UserRepository.java
@@ -4,7 +4,8 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface UserRepository extends CrudRepository<User, Integer>{
+//CRUDRepository reference http://zetcode.com/springboot/crudrepository/
+public interface UserRepository extends CrudRepository<User, Long>{
 
     @Query(
             value = "SELECT password FROM user u WHERE u.username = ?1",
